fix publish

+++ Generator.php
@@ -106,9 +106,11 @@ class FactorioGenerator
         sort($this->languages);
         $this->writeJs('translations/list', $this->languages, 'languages');
         $this->writeData();
-        unlink($this->descPath . '/index.htm');
-        unlink($this->descPath . '/index.js');
-        unlink($this->descPath . '/index.css');
+        if ($this->descPath != '.') {
+            unlink($this->descPath . '/index.htm');
+            unlink($this->descPath . '/index.js');
+            unlink($this->descPath . '/index.css');
+        }
         copy('index.htm', $this->descPath . '/index.htm');
         copy('index.js', $this->descPath . '/index.js');
         copy('index.css', $this->descPath . '/index.css');
